Move `LocalFeedImageLoader` to a separate file
